pyproject.toml
[project]
name = "kcartbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.119.0",
    "google-genai>=1.43.0",
    "langchain>=0.3.27",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "pymilvus>=2.4.0",
    "pypdf>=5.1.0",
    "requests>=2.32.3",
    "streamlit>=1.39.0",
]

[tool.aerich]
tortoise_orm = "app.core.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
